requirements.txt
psycopg2
python-dotenv
fastapi
pylint
